pyproject.toml
[project]
name = "StockAgent"
version = "0.1.0"
description = "Hierarchical Multi-Agent System for Top 5 US Stock Selection"
authors = [{ name = "Lavanya Cheshani", email = "lavanyacheshani5@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["stock", "investment", "multi-agent", "crewai", "finance", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]
dependencies = [
    "aiohttp>=3.11.13",
    "alpha-vantage>=3.0.0",
    "crewai-tools>=0.36.0",
    "crewai[tools]>=0.102.0,<1.0.0",
    "firecrawl-py>=1.6.8",
    "langsmith>=0.1.147",
    "litellm>=1.60.2",
    "lxml>=5.3.1",
    "html5lib>=1.1",
    "mkdocs>=1.6.1",
    "nest-asyncio>=1.6.0",
    "plotly>=6.0.0",
    "pydantic==2.9.2",
    "pydantic-core>=2.14.5",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.0",
    "textblob>=0.19.0",
    "vadersentiment>=3.3.2",
    "yfinance>=0.2.54",
]

[project.scripts]
stocksage = "stocksage.main:run"
run_crew = "stocksage.main:run"
train = "stocksage.main:train"
replay = "stocksage.main:replay"
test = "stocksage.main:test"

[project.urls]
Homepage = "https://github.com/Lavanyacheshani/StockAgent"
Repository = "https://github.com/Lavanyacheshani/StockAgent"
Issues = "https://github.com/Lavanyacheshani/StockAgent/issues"
Changelog = "https://github.com/Lavanyacheshani/StockAgent/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stocksage"]

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "bump-pydantic>=0.8.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests that call live external APIs (deselect with '-m not integration')",
    "slow: marks tests that take more than 30 seconds",
]
addopts = [
    "--tb=short",
    "-ra",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = [
    "B006",  # mutable default args — we know about these and are fixing them progressively
    "B008",  # function call in default arg
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # assert is fine in tests
